requirements.txt
requests
requests-oauthlib
python-dateutil
typer
python-dotenv
pydantic
tenacity
google-generativeai
google-auth
